+++ tests/run_cli_tests.sh
@@ -67,6 +67,7 @@ cat <<EOF > config_basic_test.toml
 [security]
 admin_password_bcrypt = "\$2a\$10\$Hk5/XcvwrN.JRFrjdG0vjuGZxa5JaILdir1qflIj5i9DUPUyvIK7C"
 app_default_auth_type = "system"
+auth_required = true
 
 [http]
 port = 9154